+++ CHANGELOG.rst
@@ -29,6 +29,7 @@ v2.5.0-dev
 * Add support for tabulating gene expression data to the ``hamlet_table.py`` script
 * Add support for tabulating cell types to the ``hamlet_table.py`` script
 * Add support for tabulating AML subtype prediction to the ``hamlet_table.py`` script
+* Simplify the output of the ``hamlet_table.py`` script for variants
 * Update the filter and annotation criteria to include 46 genes
 * Update MultiQC to 1.31
 * Update cutadapt from 5.0 to 5.1